Extracted the login form to its own component
